requirements.txt
asgiref==3.8.1
Django==5.1.4
django-ratelimit==4.1.0
djangorestframework==3.15.2
lxml==5.3.0
psycopg2==2.9.10
PyPDF2==3.0.1
python-docx==1.1.2
sqlparse==0.5.3
typing_extensions==4.12.2
